Changes Added the ability to swap between downloading and preparing data or just preparing data using already downloaded data
Plan
Buffer has solved the problems for the 2020, 2022, 2023 seasons but not 2021. I think this is to do with trades and injuries but still not 100% sure. For settings, I would like to test (Players_per_team = 6, Games_back = 10, Games_buffer = 4) it errors first on the portland trailblazers for game 0022101011. If I drop the number of players_per_team to 5, this will work for portland but then it errors on detroit for game 0022100540. Thinking about changing how we deal with player data. Right now it's heavily tied to teams, so if a player gets traded, they won't show up for GAMES_BACK amount of time when on a new team. I might change this to be more focused on players. This way, if a player gets traded, we can just use GAMES_BACK number of games from their old team until they have enough games on a new team. While this isn't perfect since a player role and stats can change drastically from team to team. I think it is better than the current solution as we basically ignore new acquisitions for GAMES_BACK amount of time. If that does not work for 2021 might try to use a players averages for the year when we cannot reach amount of players until we have enough stats for them