Changes Changed how we save data. We still look at seasons team by team but instead of making a team csv with all stats we instead save them to a new area /data/player/year/player_id and then combine their stats. This way if a player is traded we have their stats from before and after a trade. Also solved issue where the data would be out of order. For example if a trade happened it would be possible that the team they get traded to gets anaylzed first. Their data is in players folder, and when we append the data from the team, they got traded from which makes no sense. Data for players is now chronological
Plan
Data for players can now be used, but we still need to solve problem of knowing which players data to get. Previously this was easy since we could average the stats for the team and then just looking at the highest minutes. Now that we changed how this works, it's harder, especially since we can't do rolling averages untill all teams are done. Will have to think of a away around this and then we should almost be done with data preparation