Had issue when downloading data for individual games where we would time out. I dropped the number of threads we use, have raised the value for timeout from 30 to 120 and also added retry. Retry is used so the program will retry downloading both games and the schedule for the year set by global variable MAX_DOWNLOAD_ATTEMPTS. This has fixed the issue for years 1999-2023 and sped up the average download from 30 minutes to 10 minutes. Unsure if this is because of time I am running it or why.
Ran into new issue where games from 1990-1995 and possibly more will cause error when trying to split game data. Games have data for both teams in it, and we know which is which by checking the string value of TEAM_ABBREVIATION. For some reason, the older games are throwing the error only use .str accessor with string values!